cdk/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.119.0
aws-cdk.aws-apigateway==1.119.0
aws-cdk.aws-applicationautoscaling==1.119.0
aws-cdk.aws-autoscaling==1.119.0
aws-cdk.aws-autoscaling-common==1.119.0
aws-cdk.aws-autoscaling-hooktargets==1.119.0
aws-cdk.aws-certificatemanager==1.119.0
aws-cdk.aws-cloud9==1.119.0
aws-cdk.aws-cloudformation==1.119.0
aws-cdk.aws-cloudfront==1.119.0
aws-cdk.aws-cloudwatch==1.119.0
aws-cdk.aws-codebuild==1.119.0
aws-cdk.aws-codecommit==1.119.0
aws-cdk.aws-codeguruprofiler==1.119.0
aws-cdk.aws-codepipeline==1.119.0
aws-cdk.aws-codestarnotifications==1.119.0
aws-cdk.aws-cognito==1.119.0
aws-cdk.aws-databrew==1.119.0
aws-cdk.aws-ec2==1.119.0
aws-cdk.aws-ecr==1.119.0
aws-cdk.aws-ecr-assets==1.119.0
aws-cdk.aws-ecs==1.119.0
aws-cdk.aws-ecs-patterns==1.119.0
aws-cdk.aws-efs==1.119.0
aws-cdk.aws-elasticloadbalancing==1.119.0
aws-cdk.aws-elasticloadbalancingv2==1.119.0
aws-cdk.aws-elasticsearch==1.119.0
aws-cdk.aws-events==1.119.0
aws-cdk.aws-events-targets==1.119.0
aws-cdk.aws-globalaccelerator==1.119.0
aws-cdk.aws-glue==1.119.0
aws-cdk.aws-iam==1.119.0
aws-cdk.aws-kinesis==1.119.0
aws-cdk.aws-kinesisanalytics==1.119.0
aws-cdk.aws-kinesisfirehose==1.119.0
aws-cdk.aws-kms==1.119.0
aws-cdk.aws-lambda==1.119.0
aws-cdk.aws-logs==1.119.0
aws-cdk.aws-rds==1.119.0
aws-cdk.aws-route53==1.119.0
aws-cdk.aws-route53-targets==1.119.0
aws-cdk.aws-s3==1.119.0
aws-cdk.aws-s3-assets==1.119.0
aws-cdk.aws-s3-deployment==1.119.0
aws-cdk.aws-sam==1.119.0
aws-cdk.aws-secretsmanager==1.119.0
aws-cdk.aws-servicediscovery==1.119.0
aws-cdk.aws-signer==1.119.0
aws-cdk.aws-sns==1.119.0
aws-cdk.aws-sns-subscriptions==1.119.0
aws-cdk.aws-sqs==1.119.0
aws-cdk.aws-ssm==1.119.0
aws-cdk.aws-stepfunctions==1.119.0
aws-cdk.cloud-assembly-schema==1.119.0
aws-cdk.core==1.119.0
aws-cdk.custom-resources==1.119.0
aws-cdk.cx-api==1.119.0
aws-cdk.lambda-layer-awscli==1.119.0
aws-cdk.region-info==1.119.0
cattrs==1.6.0
certifi==2021.5.30
charset-normalizer==2.0.6
constructs==3.3.119
idna==3.2
jsii==1.32.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.7
